+++ src/main/java/xpto/repository/PartnersRepository.java
@@ -3,5 +3,5 @@ package xpto.repository;
 import xpto.entity.Partner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartnersRepository extends JpaRepository<Partner, Long> {
+public interface PartnersRepository extends JpaRepository<Partner, Integer> {
 }
