+++ src/main/java/xpto/repository/PersonRepository.java
@@ -3,8 +3,8 @@ package xpto.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import xpto.entity.Person;
 
-public interface PersonRepository extends JpaRepository<Person, Long> {
+public interface PersonRepository extends JpaRepository<Person, Integer> {
 
-    Person findByIdAndPersonType(Long id, String personType);
+    Person findByIdAndPersonType(int id, String personType);
 
 }
